requirements.txt
kubernetes>=32.0.0
PyYAML>=6.0